Move GameBoy serialisers into their own folder.
