+++ wwwroot/modules/serialisers/gameBoy/gameBoyTileBinarySerialiser.js
@@ -1,5 +1,5 @@
-import Tile from '../models/tile.js';
-import TileFactory from '../factory/tileFactory.js';
+import Tile from '../../models/tile.js';
+import TileFactory from '../../factory/tileFactory.js';
 
 
 export default class GameBoyTileBinarySerialiser {
