+++ wwwroot/modules/serialisers/projectAssemblySerialiser.js
@@ -4,8 +4,8 @@ import ColourUtil from "../util/colourUtil.js";
 import Project from "../models/project.js";
 import TileSetBinarySerialiser from "./tileSetBinarySerialiser.js";
 import TileMapBinarySerialiser from "./tileMapBinarySerialiser.js";
-import GameBoyTileSetBinarySerialiser from "./gameBoyTileSetBinarySerialiser.js";
-import GameBoyTileMapBinarySerialiser from "./gameBoyTileMapBinarySerialiser.js";
+import GameBoyTileSetBinarySerialiser from "./gameBoy/gameBoyTileSetBinarySerialiser.js";
+import GameBoyTileMapBinarySerialiser from "./gameBoy/gameBoyTileMapBinarySerialiser.js";
 import TileMapUtil from "../util/tileMapUtil.js";
 import TileSetFactory from "../factory/tileSetFactory.js";
 import TileMap from "../models/tileMap.js";
